tester.py: Implement protocol testing
The tests.c file and tester.py implement protocol testing for MAC, IPv6, UDP
and COAP protocols.

The tests.c implements tests on the device side, while the tester.py script
implements test execution. The tester.py script relies on scapy to build and
decode frames. The list of tests to execute is selected in the tester.py
script.

Signed-off-by: Emmanuel Thierry <emmanuel.thierry@sekil.fr>